requirements.txt
asgiref==3.3.4
astroid==2.4.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.16
django-environ==0.4.5
djangorestframework==3.12.4
dodgy==0.2.1
drf-yasg==1.20.0
flake8==3.8.4
flake8-polyfill==1.0.2
idna==2.10
importlib-metadata==4.0.0
inflection==0.5.1
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.9
pep8-naming==0.10.0
prospector==1.3.1
psycopg2==2.8.6
pycodestyle==2.6.0
pydocstyle==6.0.0
pyflakes==2.2.0
pylint==2.5.3
pylint-celery==0.3
pylint-django==2.1.0
pylint-flask==0.6
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requirements-detector==0.7
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
setoptconf==0.2.0
six==1.15.0
snowballstemmer==2.1.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
wrapt==1.12.1
zipp==3.4.1
